requirements.txt
numpy==1.20.3
pillow==8.4.0